# ld-service/hlahapv-swagger-spec.yaml
swagger: '2.0'
info:
  description: |
    This is a HLAHapV Server.  You can find 
    out more about HLAHapV at 
    [http://github.com/nmdp-bioinformatics/ImmunogeneticDataTools](http://github.com/nmdp-bioinformatics/ImmunogeneticDataTools).
  version: 1.0.0
  title: Swagger HLAHapV
  termsOfService: http://swagger.io/terms/
  contact:
    email: mpresteg@nmdp.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# basePath: /v2
tags:
- name: genotypes
  description: Genotypes to be evaluated
# schemes:
# - http
paths:
  /genotypes:
    post:
      tags:
      - genotypes
      summary: Submit a set of genotypes for evaluation
      operationId: submitGenotypes
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: Genotypes object that needs to be evaluated
        required: true
        schema:
          $ref: '#/definitions/Genotypes'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Samples'
        405:
          description: Invalid input
  /genotypes/file:
    post:
      tags:
      - genotypes
      summary: Submit a batch file of genotypes for evaluation (the same tab-delimited
        id/GL-String-per-line format ld-tools' analyze-gl-strings reads)
      operationId: submitGenotypesFile
      consumes:
      - multipart/form-data
      produces:
      - application/json
      - application/xml
      parameters:
      - in: formData
        name: file
        description: batch input file, one "id<TAB>glString" pair per line
        required: true
        type: file
      - in: formData
        name: hladbVersion
        description: IMGT/HLA reference database version to validate against (e.g.
          3.19.0), default latest
        required: false
        type: string
      - in: formData
        name: frequencySet
        description: named frequency set to use (e.g. nmdp-2007, nmdp-2007-std,
          wiki), default nmdp-2007-std. Ignored if frequencyFiles is supplied.
          Note that "nmdp" (as opposed to "nmdp-2007"/"nmdp-2007-std") has no
          bundled reference data and will fail -- it's reserved for a
          licensed dataset this project never redistributes.
        required: false
        type: string
      - in: formData
        name: frequencyFiles
        description: your own frequency reference file(s), in the "standard
          format" ld-tools' normalize-frequency-file produces (from NMDP data or
          any other source) -- the same files the CLI's -q/--frequency-file(s)
          flag takes. Overrides frequencySet when present.
        required: false
        type: array
        items:
          type: file
        collectionFormat: multi
      - in: formData
        name: allelesFile
        description: optional individual-locus allele frequency file accompanying
          frequencyFiles (the CLI's -l/--allele-file). Ignored if frequencyFiles
          is not also supplied.
        required: false
        type: file
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Samples'
        405:
          description: Invalid input
definitions:
  Genotypes:
    type: object
    required:
    - genotype
    properties:
      genotype:
        type: array
        xml:
          name: genotype
          wrapped: true
        items:
          $ref: '#/definitions/Genotype'
      hladbVersion:
        type: string
        description: IMGT/HLA reference database version to validate against (e.g.
          3.19.0), default latest. Applies to the whole request, not per genotype.
      frequencySet:
        type: string
        description: named frequency set to use (e.g. nmdp-2007, nmdp-2007-std,
          wiki), default nmdp-2007-std. Applies to the whole request, not per
          genotype. Note that "nmdp" (as opposed to "nmdp-2007"/"nmdp-2007-std")
          has no bundled reference data and will fail -- it's reserved for a
          licensed dataset this project never redistributes. To use your own
          frequency file(s) instead of a named set, use POST /genotypes/file's
          frequencyFiles field -- this JSON endpoint can't carry file uploads.
    xml:
      name: genotypes
  Genotype:
    type: object
    required:
    - id
    - glString
    properties:
      id:
        type: string
      glString:
        type: string
    xml:
      name: genotype
  Samples:
    type: object
    required:
    - sampleData
    properties:
      sample:
        type: array
        xml:
          name: sampleData
          wrapped: true
        items:
          $ref: '#/definitions/SampleData'
    xml:
      name: samples  
  SampleData:
    type: object
    required:
    - id
    - glString
    properties:
      id:  
        type: string
      glString:
        type: string
      processedGlString:
        type: string
      haplotypePair:
        type: array
        xml:
          name: haplotypePair
          wrapped: true
        items:
          $ref: '#/definitions/HaplotypePairData'
      hasAnomalies:
        type: boolean
        description: true if detection found no linkages, no haplotype pairs for
          a sought locus combination, non-CWD/non-CIWD alleles, or a G-group/ARS
          match against more than the expected reference entries -- the same "should
          this get a human's attention" signal analyze-gl-strings uses to route
          a sample to its *Warning*.log files instead of the plain ones.
      warnings:
        type: array
        description: human-readable anomaly summaries (DetectedLinkageFindings#getWarnings)
        items:
          type: string
      linkageReport:
        type: string
        description: same text analyze-gl-strings writes to linkages.log/linkageWarnings.log
          for this sample (LinkageDisequilibriumWriter#formatDetectedLinkages)
      haplotypePairReport:
        type: string
        description: same text analyze-gl-strings writes to haplotypePairs.log/haplotypePairsWarning.log
          for this sample (HaplotypePairWriter#formatDetectedLinkages)
      nonCommonWellDocumentedReport:
        type: string
        description: same text analyze-gl-strings writes to nonCwdWarnings.log for
          this sample (CommonWellDocumentedWriter#formatCommonWellDocumented); empty
          when there are no non-CWD/non-CIWD alleles
      detectedFindingsReport:
        type: string
        description: same CSV row(s) analyze-gl-strings writes to detectedFindings.csv
          for this sample (DetectedFindingsWriter#formatDetectedFindings)
    xml:
      name: sample
  FindingData:
    type: object
    properties:
      frequency:
        type: number
      relativeFrequency:
        type: number
      haplotype1Frequency:
        type: number
      haplotype2Frequency:
        type: number
      race:
        type: string
    xml:
      name: finding 
  HaplotypePairData:
    type: object
    properties:
      haplotype1:
        type: string
      haplotype2:
        type: string
      finding:
        type:  array
        xml:
          name: finding
          wrapped: true
        items:
          $ref: '#/definitions/FindingData'
            
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /mpresteg/HLAHapV/1.0.0
schemes:
 - https